[video_player_avplay] Fix start position out of range issue (#773)

+++ packages/video_player_avplay/tizen/src/plus_player.cc
@@ -138,7 +138,7 @@ int64_t PlusPlayer::Create(const std::string &uri,
   }
 
   int64_t start_position = flutter_common::GetValue(
-      create_message.player_options(), "startPosition", 0);
+      create_message.player_options(), "startPosition", (int64_t)0);
   if (start_position > 0) {
     LOG_INFO("[PlusPlayer] Start position: %lld", start_position);
     if (!Seek(player_, start_position)) {
